requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
aiosqlite==0.20.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chatgpt_md_converter==0.1.2
cryptography==42.0.8
emoji==2.12.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
langdetect==1.0.9
mistletoe==1.3.0
multidict==6.0.5
mysql-connector-python==9.0.0
pillow==10.4.0
pycparser==2.22
pyTelegramBotAPI==4.20.0
python-dotenv==1.0.1
python-telegram-bot==21.3
requests==2.32.3
six==1.16.0
sniffio==1.3.1
telegramify-markdown==0.1.8
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
alibabacloud_green20220302==2.2.11
oss2==2.18.6